backend/requirements.txt
fastapi
uvicorn[standard]
openai
google-generativeai
anthropic
pyyaml
pydantic
python-dotenv

